combine auth buttons
