+++ src/styles/header.scss
@@ -52,7 +52,8 @@ header {
       font-size: 14px;
       padding-right: 1.2em;
       padding-left: 1.2em;
-      display: inline-block;
+      display: flex;
+      align-items: center;
       position: relative;
       &.active {
         background-color: $primaryProgressBtnColor;
